+++ packages/ovm/src/contracts/DataTypes.sol
@@ -26,6 +26,6 @@ contract DataTypes {
         address ovmActiveContract;
         address ovmMsgSender;
         address ovmTxOrigin;
-        address l1TxSender;
+        address l1MessageSender;
     }
 }
